scripts/requirements-sdxl.txt
torch>=2.2,<2.6
diffusers>=0.30,<0.33
transformers>=4.40,<4.48
accelerate>=0.30,<1.2
safetensors>=0.4,<0.6
Pillow>=10,<12
huggingface-hub>=0.24,<0.28

